Add workaround for blur in `ElementLogoAtom` for OS < 9.
