+++ libraries/designsystem/src/main/kotlin/io/element/android/libraries/designsystem/atomic/atoms/ElementLogoAtom.kt
@@ -16,7 +16,6 @@
 
 package io.element.android.libraries.designsystem.atomic.atoms
 
-import android.graphics.BlurMaskFilter
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -27,24 +26,16 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.blur
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.CornerRadius
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Rect
-import androidx.compose.ui.geometry.RoundRect
-import androidx.compose.ui.graphics.ClipOp
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Paint
-import androidx.compose.ui.graphics.Path
-import androidx.compose.ui.graphics.drawscope.clipPath
-import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
-import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import io.element.android.libraries.designsystem.R
+import io.element.android.libraries.designsystem.modifiers.blurCompat
+import io.element.android.libraries.designsystem.modifiers.blurredShapeShadow
+import io.element.android.libraries.designsystem.modifiers.canUseBlurMaskFilter
 import io.element.android.libraries.designsystem.preview.DayNightPreviews
 import io.element.android.libraries.designsystem.preview.ElementPreview
 import io.element.android.libraries.theme.ElementTheme
@@ -66,22 +57,35 @@ fun ElementLogoAtom(
             .border(size.borderWidth, borderColor, RoundedCornerShape(size.cornerRadius)),
         contentAlignment = Alignment.Center,
     ) {
-        Box(
-            Modifier
-                .size(size.outerSize)
-                .shapeShadow(
-                    color = shadowColor,
-                    cornerRadius = size.cornerRadius,
-                    blurRadius = size.shadowRadius,
-                    offsetY = 8.dp,
-                )
-        )
+        if (canUseBlurMaskFilter()) {
+            Box(
+                Modifier
+                    .size(size.outerSize)
+                    .blurredShapeShadow(
+                        color = shadowColor,
+                        cornerRadius = size.cornerRadius,
+                        blurRadius = size.shadowRadius,
+                        offsetY = 8.dp,
+                    )
+            )
+        } else {
+            Box(
+                Modifier
+                    .size(size.outerSize)
+                    .shadow(
+                        elevation = size.shadowRadius,
+                        shape = RoundedCornerShape(size.cornerRadius),
+                        clip = false,
+                        ambientColor = shadowColor
+                    )
+            )
+        }
         Box(
             Modifier
                 .clip(RoundedCornerShape(size.cornerRadius))
                 .size(size.outerSize)
                 .background(backgroundColor)
-                .blur(blur)
+                .blurCompat(blur)
         )
         Image(
             modifier = Modifier.size(size.logoSize),
@@ -120,44 +124,6 @@ sealed class ElementLogoAtomSize(
         shadowRadius = 60.dp,
     )
 }
-
-fun Modifier.shapeShadow(
-    color: Color = Color.Black,
-    cornerRadius: Dp = 0.dp,
-    offsetX: Dp = 0.dp,
-    offsetY: Dp = 0.dp,
-    blurRadius: Dp = 0.dp,
-) = then(
-    drawBehind {
-        drawIntoCanvas { canvas ->
-            val path = Path().apply {
-                addRoundRect(RoundRect(Rect(Offset.Zero, size), CornerRadius(cornerRadius.toPx())))
-            }
-
-            clipPath(path, ClipOp.Difference) {
-                val paint = Paint()
-                val frameworkPaint = paint.asFrameworkPaint()
-                if (blurRadius != 0.dp) {
-                    frameworkPaint.maskFilter = BlurMaskFilter(blurRadius.toPx(), BlurMaskFilter.Blur.NORMAL)
-                }
-                frameworkPaint.color = color.toArgb()
-
-                val leftPixel = offsetX.toPx()
-                val topPixel = offsetY.toPx()
-                val rightPixel = size.width + topPixel
-                val bottomPixel = size.height + leftPixel
-
-                canvas.drawRect(
-                    left = leftPixel,
-                    top = topPixel,
-                    right = rightPixel,
-                    bottom = bottomPixel,
-                    paint = paint,
-                )
-            }
-        }
-    }
-)
 
 @Composable
 @DayNightPreviews
